Admin aimi

+++ database/migrations/2026_05_25_090000_create_review_images_table.php
@@ -9,17 +9,20 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
+   public function up(): void
     {
-        Schema::create('review_images', function (Blueprint $table) {
-            $table->id();
-
-            $table->foreignId('review_id')->constrained()->onDelete('cascade');
-
-            // store the path of the image in the storage
-            $table->string('image_path');
-            $table->timestamps();
-        });
+        // 💡 もし 'review_images' テーブルが存在「しない」場合だけ作成する処理
+        if (!Schema::hasTable('review_images')) {
+            Schema::create('review_images', function (Blueprint $table) {
+                $table->id();
+                
+                $table->foreignId('review_id')->constrained()->onDelete('cascade');
+                
+                // store the path of the image in the storage
+                $table->string('image_path');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
